pyproject.toml
[tool.poetry]
name = "pg-budget"
version = "0.1.0"
description = ""
authors = ["Gael <prungael@gmail.com>"]
readme = "README.md"

packages = [
    { include = "pg_budget" }
]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
python-dateutil = "^2.9.0.post0"
pyside6 = "^6.9.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-qt = "^4.5.0"
pytest-mock = "^3.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pg-budget = "pg_budget.main:main"